Refactor performance benchmarking to explicitly track operation success
- Updated `_time_operation` method to return a success flag alongside the result and elapsed time.
- Adjusted benchmarking logic in `PerformanceBenchmarker` to handle success and failure cases for fitting and predicting models.
- Enhanced error reporting in benchmark results to include error messages when operations fail.

Improve Python demo for model selection with explicit dtype handling

- Modified `train_test_split` in the Python demo to ensure `y` is of type `float64` to match PyO3 requirements.

Implement KMeans and DBSCAN clustering with enhanced functionality

- Refactored `PyKMeans` and `PyDBSCAN` to include additional parameters and methods for fitting, predicting, and retrieving fitted attributes.
- Added core logic for fitting and predicting to allow for unit testing without a live Python interpreter.

Update library versioning and dependencies

- Changed version reporting in the library to use the package version from `Cargo.toml`.
- Updated dependencies in `Cargo.toml` files across various crates to use workspace features for better management.

Enhance classification metrics with improved error handling

- Added checks for contiguous arrays in classification metrics functions to prevent internal errors.
- Updated validation logic to ensure input arrays have the same length before processing.

Refactor model selection utilities for better compatibility

- Implemented core logic for `train_test_split` and `KFold` to allow for unit testing and improved error handling.
- Ensured that `KFold` correctly handles the number of splits and provides appropriate error messages for invalid configurations.

+++ crates/sklears-python/examples/python_demo.py
@@ -121,8 +121,13 @@ def demo_model_selection():
     # Generate sample data
     X, y = datasets.make_classification(n_samples=200, n_features=10, n_classes=2, random_state=42)
 
-    # Train-test split
-    X_train, X_test, y_train, y_test = skl.train_test_split(X, y, test_size=0.3, random_state=42)
+    # Train-test split. train_test_split's PyO3 signature requires exact-dtype
+    # PyReadonlyArray1<f64> for y; make_classification() returns int64 labels
+    # on 64-bit platforms, so the dtype must be made explicit here (same
+    # reasoning as the accuracy_score dtype note in demo_metrics() below).
+    X_train, X_test, y_train, y_test = skl.train_test_split(
+        X, y.astype(np.float64), test_size=0.3, random_state=42
+    )
 
     print(f"Dataset split - Train: {X_train.shape[0]}, Test: {X_test.shape[0]}")
